requirements.txt
torch
torchvision
pillow
datasets
tqdm
matplotlib